Add reducers folder and repack youtubeSearchCacheReducer file
